Add some tests for /overview handler
* Add various test cases:
  * Single record from a player on a map
  * No record on a map
  * /overview of top 15 players of a map
  * /overview of players who don't have a record on a map
  * /overview of players not in top 15
  * /overview with records having same time (showing competition
    ranking)
* Thanks to the test made during implementation, fix some bugs with the
  /overview request
* Add cfg flag in build.rs to automatically delete the database after a
  test, even when it fails
* Catch panic from tests in `base::wrap` function

+++ crates/admin/src/leaderboard.rs
@@ -10,10 +10,10 @@ pub enum LbCommand {
 #[clap(name = "full")]
 pub struct FullCmd {
     #[arg(long)]
-    offset: Option<i64>,
+    offset: Option<i32>,
 
     #[arg(long, short = 'n')]
-    limit: Option<i64>,
+    limit: Option<i32>,
 
     #[clap(subcommand)]
     map: Map,
@@ -36,8 +36,8 @@ async fn mariadb_lb<C: ConnectionTrait + StreamTrait>(
     conn: &C,
     redis_conn: &mut RedisConnection,
     map_id: u32,
-    offset: Option<i64>,
-    limit: Option<i64>,
+    offset: Option<i32>,
+    limit: Option<i32>,
 ) -> anyhow::Result<()> {
     let leaderboard =
         leaderboard::leaderboard(conn, redis_conn, map_id, offset, limit, Default::default())
